Renaming juniper and arista files
