+++ netmiko/arista/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import unicode_literals
-from netmiko.arista.arista_ssh import AristaSSH, AristaTelnet, AristaFileTransfer
+from netmiko.arista.arista import AristaSSH, AristaTelnet, AristaFileTransfer
 
 __all__ = ['AristaSSH', 'AristaTelnet', 'AristaFileTransfer']
